services/image_service.py
"""
    This script demonstrates how to resize an image using the Python Imaging Library (PIL).
"""
from PIL import Image

# Open an image file
with Image.open('example.jpg') as img:
    # Resize the image
    img_resized = img.resize((800, 600))
    # Save the resized image
    img_resized.save('example_resized.jpg')


# Upload Image: Allow users to upload images.
def authorized():
    pass


def upload_image(image):
    if not authorized():
        return 'Unauthorized', 401
    
# Transform Image: Allow users to perform various transformations (resize, crop, rotate, watermark etc.).

# Retrieve Image: Allow users to retrieve a saved image in different formats.

# List Images: List all uploaded images by the user with metadata.
