requirements.txt
flask
pylint
pytest
Pillow
python-dotenv